% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compas.R
\name{compas}
\alias{betapas}
\alias{compas}
\title{Community Pattern Simulation}
\usage{
compas(nsp, n, xgrad = 6, ygrad = 4, x, xn = 6, yn = 4, buffer = 2,
  tsd = 0.1)

betapas(nsp, n, xgrad = 6, ygrad = 4, x, xn = 6, yn = 4, buffer = 2,
  tsd = 0.1, range = 5, shape = c(0.5, 6.5))
}
\arguments{
\item{nsp}{Number of simulated species. Some species will not occur
in any simulated sample and the number of species is usually lower
in simulation results.}

\item{n}{Number of sample plots.}

\item{xgrad,ygrad}{Lengths of  gradients in \eqn{t} units
(a.k.a. \sQuote{sd} units).}

\item{x}{Coordinates for gradient positions of sample plots.}

\item{xn,yn}{Number of sample plots along gradients in a
regular grid. This argument will be used only if \code{x} was not
given.}

\item{buffer}{Buffer zone around the sampled gradient space for
simulated species optima.}

\item{tsd}{Standard deviation of widths of species responses in
lognormal distribution.}

\item{range}{The range parameter of beta response.}

\item{shape}{The range of shape parameters (\eqn{alpha},
\eqn{gamma}) in uniform distribution.}
}
\description{
Functions generate simulated binary occurrence data from Gaussian
response (\code{compas}) or beta response functions
(\code{betapas}).
}
\details{
Function \code{compas} generates simulated data from
Gaussian species response function with Bernoulli error. The
species optima are distributed uniformly in the gradient space
(plus \code{buffer}), the response heights are distributed
uniformly in \eqn{(0,1)}, and the response widths are lognormally
distributed with mean \eqn{t=1} and standard deviation as given in
the function call.

Function \code{betapas} generates data with Bernoulli error from
beta response function. The function first generates Gaussian
species parameters as in \code{compas}, and then transforms these
to beta responses with the same optima and average heights
(occurrence probability) as the corresponding Gaussian responses,
but with varying shapes (skewness, kurtosis).
}
\section{Functions}{
\itemize{
\item \code{compas}: Simulation with Gaussian response functions.

\item \code{betapas}: Simulation with beta response functions.
}}

