% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GO.R
\name{GO1}
\alias{GO1}
\alias{GO2}
\alias{anova.GO2}
\alias{calibrate.GO2}
\alias{plot.GO1}
\alias{plot.GO2}
\alias{predict.GO2}
\alias{spanodev}
\title{Unconstrained Gaussian Maximum Likelihood Ordination}
\usage{
GO1(comm, tot = max(comm), freqlim = 5, parallel = 1, ...)

GO2(comm, k = 1, tot = max(comm), freqlim = 5, family = c("poisson",
  "binomial"), far = 10, ...)

\method{plot}{GO1}(mod, ...)

\method{plot}{GO2}(mod, choices = 1, label = FALSE, marginal = FALSE,
  cex = 0.7, col = 1:6, ...)

\method{anova}{GO2}(object, ...)

spanodev(mod1, mod2 = NULL, ...)

\method{predict}{GO2}(object, newdata, type = c("response", "link"), ...)

calibrate.GO2(object, newdata, ...)
}
\arguments{
\item{comm}{Community data matrix of data frame.}

\item{tot}{Maximum abundance value or total used in Binomial models.}

\item{freqlim}{Minimum number of occurrence for analysed species.}

\item{parallel}{Number of parallel processes.}

\item{k}{Number of estimated gradients (axes).}

\item{family}{Error distribution.}

\item{far}{Threshold distance for species optima regarded as
outliers and frozen in fitting.}

\item{mod}{Fitted model.}

\item{choices}{The axis or axes plotted.}

\item{label}{Label species responses.}

\item{marginal}{Plot marginal responses or slice through origin of
other dimensions.}

\item{cex}{Character size for \code{labels}.}

\item{col}{Colours of response curves.}

\item{object}{Ordination result object.}

\item{mod1,mod2}{Compared result objects}

\item{newdata}{New gradient locations for predicted responses}

\item{type}{Predictions in the scale of responses or in the scale
of link function.}

\item{\dots}{Other parameters passed to functions.}
}
\description{
Unconstrained Gaussian Maximum Likelihood Ordination
}
\section{Functions}{
\itemize{
\item \code{GO1}: Alternating estimation of species parameters and
gradient locations in one dimension.

\item \code{GO2}: Simultaneous estimation of species parameters and
gradient locations.

\item \code{spanodev}: Comparison of goodness of fit for individual species.
}}

