% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simul.R
\name{simulrun}
\alias{simulrun}
\title{Run a Set of Simulations and Collect Results}
\usage{
simulrun(nsimul = 1, nsp = 300, n = 100, xgrad = 2, ygrad = 2,
  respfun = compas, ...)
}
\arguments{
\item{nsimul}{Number of simulations.}

\item{nsp}{Number of species in simulations passed to
\code{respfun}. The realized number can be (and usually is) lower
because some species do not occur in any simulated communities.}

\item{n}{Number of sample plots passed to \code{respfun}.}

\item{xgrad,ygrad}{Lengths of gradients passed to \code{respfun}.}

\item{respfun}{The response function used in simulations.}

\item{\dots}{Other arguments passed to \code{respfun}.}
}
\description{
Function is a general driver to run several ordination methods on a
set of simulated data sets and collect the goodness of fit
statistics for the results.
}

