% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coenoclinerutil.R
\name{GradLocs}
\alias{BinomGaussPar}
\alias{DropMissingSpec}
\alias{Gauss2betaPar}
\alias{GradLocs}
\alias{coenorun1}
\title{Utility Functions for 'coenocliner' Package}
\usage{
GradLocs(n, xrange, yrange)

BinomGaussPar(nsp, xrange, yrange, buffer = 2, tsd = 0.1)

Gauss2betaPar(gausspar, shape = c(0.5, 6.5), range = 5)

DropMissingSpec(comm)

coenorun1(sim, locs, tot = 1)
}
\arguments{
\item{n}{Number of SUs.}

\item{xrange,yrange}{Desired range of gradients.}

\item{nsp}{Number of species.}

\item{buffer}{Width of buffer zone for optima surrounding ranges.}

\item{tsd}{Standard deviation of tolerance in log-Normal
distribution, in log scale}

\item{gausspar}{Gaussian response parameters for species as
returned by \code{BinomGaussPar}.}

\item{shape}{Random log-uniform range of shape parameters \eqn{alpha}
and \eqn{gamma} of response function}

\item{range}{Range of beta response in $t$ (\sQuote{sd}) units of
Gaussian response function.}

\item{comm}{Community data.}

\item{sim}{One simulated community.}

\item{locs}{True gradient locations.}

\item{tot}{Binomial total in \code{sim}.}
}
\description{
Functions to automated simulation routines using \pkg{coenocliner}
package.


}
\section{Functions}{
\itemize{
\item \code{GradLocs}: Gradient Locations

\item \code{BinomGaussPar}: Gaussian Parameters for Binomial Response.

\item \code{Gauss2betaPar}: Translate Gaussian parameters into
corresponding beta response parameters.

\item \code{DropMissingSpec}: Drop missing species from the data.

\item \code{coenorun1}: Takes one simulated community for
ordination with GO, NMDS, CA and DCA and returns average Procrustes
precision
}}
\examples{
if (require(coenocliner)) {
gp <- BinomGaussPar(400, 6, 2)
bp <- Gauss2betaPar(gp)
xy62 <- GradLocs(100, 6, 2)
xy31 <- GradLocs(100, 3, 1)
coenorun1(coenocline(xy31, "gaussian", gp, countModel="bernoulli"), xy62)
coenorun1(coenocline(xy31, "gaussian", gp, countModel="bernoulli"), xy31)
coenorun1(coenocline(xy31, "beta", bp, countModel="bernoulli"), xy62)
coenorun1(coenocline(xy31, "beta", bp, countModel="bernoulli"), xy31)
}
}

