% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coenoclinerutil.R
\name{GradLocs}
\alias{BinomGaussPar}
\alias{DropMissingSpec}
\alias{Gauss2betaPar}
\alias{GradLocs}
\alias{GradMul}
\alias{coenorun1}
\title{Utility Functions for 'coenocliner' Package}
\usage{
GradLocs(n, xrange, yrange)

GradMul(xy, xmul, ymul)

BinomGaussPar(nsp, xrange, yrange, buffer = 2, tsd = 0.1)

Gauss2betaPar(gausspar, shape = c(0.5, 6.5), cover = 0.95)

DropMissingSpec(comm)

coenorun1(sim, tot = 1, family = "binomial", far = 4, trace = TRUE)
}
\arguments{
\item{n}{Number of SUs.}

\item{xrange,yrange}{Desired range of gradients.}

\item{xy}{Gradient locations in two dimensions.}

\item{xmul,ymul}{Multipliers for each gradient}

\item{nsp}{Number of species.}

\item{buffer}{Width of buffer zone for optima surrounding ranges.}

\item{tsd}{Standard deviation of tolerance in log-Normal
distribution, in log scale}

\item{gausspar}{Gaussian response parameters for species as
returned by \code{BinomGaussPar}.}

\item{shape}{Random log-uniform range of shape parameters \eqn{alpha}
and \eqn{gamma} of response function}

\item{cover}{Find range of beta response so that the same span
covers the same proportion of 1-dim integral as the Gaussian
response function.}

\item{comm}{Community data.}

\item{sim}{One simulated community.}

\item{tot}{Binomial total in \code{sim}.}

\item{family}{Error family passed to \code{\link{GO}}.}

\item{far}{Weirdness limit passed to \code{\link{GO}}.}

\item{trace}{Print tracing information. If \code{FALSE} or
\code{0}, work as silently as possible, and higher values print more.}
}
\description{
Functions to automated simulation routines using \pkg{coenocliner}
package.


}
\section{Functions}{
\itemize{
\item \code{GradLocs}: Gradient Locations

\item \code{GradMul}: Multiply input gradient which
presumably is a unit square

\item \code{BinomGaussPar}: Gaussian Parameters for Binomial Response.

\item \code{Gauss2betaPar}: Translate Gaussian parameters into
corresponding beta response parameters.

\item \code{DropMissingSpec}: Drop missing species from the data.

\item \code{coenorun1}: Takes one simulated community for
ordination with GO, NMDS, CA and DCA and returns average Procrustes
precision
}}
\examples{
require(coenocliner) || stop("examples need 'coenocliner' package")
## small simulation
nsim <- 10
npoints <- 50
## generate a set of species parameters over the maximum extent
sp <- replicate(nsim, BinomGaussPar(800, 8, 4))
## sample much narrower proportion of the space
xy <- replicate(nsim, GradLocs(npoints, 3, 2))
## Simulations: these can be easily parallelized using mclapply
## (Linux, Mac) or parSapply (all).
sapply(seq_len(nsim), function(i)
     coenorun1(coenocline(xy[,,i], "gaussian", sp[,i],
         countModel="bernoulli")))
}
\author{
Jari Oksanen
}

