% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coenoclinerutil.R
\name{GradLocs}
\alias{BinomGaussPar}
\alias{DropMissingSpec}
\alias{GradLocs}
\title{Utility Functions for 'coenocliner' Package}
\usage{
GradLocs(n, xrange, yrange)

BinomGaussPar(nsp, xrange, yrange, buffer = 2, tsd = 0.1)

DropMissingSpec(comm)
}
\arguments{
\item{n}{Number of SUs.}

\item{xrange,yrange}{Desired range of gradients.}

\item{nsp}{Number of species.}

\item{buffer}{Width of buffer zone for optima surrounding ranges.}

\item{tsd}{Standard deviation of tolerance in log-Normal
distribution, in log scale}

\item{comm}{Community data.}
}
\description{
Functions to automated simulation routines using \pkg{coenocliner}
package.


}
\section{Functions}{
\itemize{
\item \code{GradLocs}: Gradient Locations

\item \code{BinomGaussPar}: Gaussian Parameters for Binomial Response.

\item \code{DropMissingSpec}: Drop missing species from the data.
}}

