% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GO.R
\name{GO1}
\alias{GO1}
\alias{GO2}
\alias{anova.GO}
\alias{calibrate.GO}
\alias{plot.GO}
\alias{predict.GO}
\alias{spanodev}
\title{Unconstrained Gaussian Maximum Likelihood Ordination}
\usage{
GO1(comm, tot = max(comm), freqlim = 5, parallel = 1, ...)

GO2(comm, k = 1, tot = max(comm), freqlim = 5, family = c("poisson",
  "binomial"), far = 10, init, ...)

\method{plot}{GO}(x, choices = 1, label = FALSE, marginal = FALSE,
  cex = 0.7, col = 1:6, ...)

\method{anova}{GO}(object, ...)

spanodev(mod1, mod2 = NULL, ...)

\method{predict}{GO}(object, newdata, type = c("response", "link"), ...)

\method{calibrate}{GO}(object, newdata, ...)
}
\arguments{
\item{comm}{Community data frame.}

\item{tot}{Total abundance used in Binomial models. This can be
either a single value for all data, or a vector with value for each
row of \code{comm}. The default is to use the maximum value in
matrix.}

\item{freqlim}{Minimum number of occurrence for analysed species.}

\item{parallel}{Number of parallel processes.}

\item{k}{Number of estimated gradients (axes).}

\item{family}{Error distribution. Can be either \code{"poisson"}
for quasi-Poisson or \code{"binomial"} for quasi-Binomial (and must
be quoted).}

\item{far}{Threshold distance for species optima regarded as
alien and frozen in fitting.}

\item{init}{Initial configuration to start the iterations. This
should be a matrix, and number of rows should match the community
data, and number coluns the number of gradients (\code{k}). The
default is to use scores from \code{\link[vegan]{decorana}}.}

\item{x}{Fitted model.}

\item{choices}{The axis or axes plotted.}

\item{label}{Label species responses.}

\item{marginal}{Plot marginal responses or slice through origin of
other dimensions.}

\item{cex}{Character size for \code{labels}.}

\item{col}{Colours of response curves.}

\item{object}{Ordination result object.}

\item{mod1,mod2}{Compared result objects}

\item{newdata}{New gradient locations to \code{predict} species
responses or new community data to \code{calibrate} gradient
locations.}

\item{type}{Predictions in the scale of responses or in the scale
of link function.}

\item{\dots}{Other parameters passed to functions. In \code{GO2}
these are passed to \code{\link{nlm}} and can include, e.g.,
\code{iterlim} (which often must be set to higher value than the
default 100).}
}
\description{
The functions fit unconstrained maximum likelihood ordination with
unit-width Gaussian response models.
}
\details{
Function is under development and unreleased. It will be released
 under different name in \pkg{vegan}. The current version is only
 provided for review purposes. The function and its support functions
 require \pkg{vegan}, although this requirements is not explicitly
 made. The optimization is based on \code{\link{nlm}} function, and passes
 arguments to this function.

 Function \code{anova} can analyse two nested models or a single model
 against null model of flat responses using parametric tests based on
 quasi-Likelihood. Function \code{spanodev} performs similar test
 split by species. Function \code{predict} returns estimated response
 curves, and \code{newdata} can be gradient locations. Function
 \code{calibrate} returns estimated gradient locations, and \code{newdata}
 can be community data.

 The \code{plot} function displays fitted respose curves against one
 ordination axis. In principle, the ordination can be rotated using
 \pkg{vegan} function \code{\link[vegan]{MDSrotate}}, but this requires
 a version that agrees to analyse \code{GO2} results. Traditional
 ordination plots of SU scores and species optima can be displayed
 with \code{\link[vegan]{ordiplot}} (\pkg{vegan} package). The function
 is written so that several other \pkg{vegan} and standard \R functions
 can handle results.
}
\section{Functions}{
\itemize{
\item \code{GO1}: Alternating estimation of species parameters and
gradient locations in one dimension.

\item \code{GO2}: Simultaneous estimation of species parameters and
gradient locations.

\item \code{spanodev}: Comparison of goodness of fit for individual species.
}}
\examples{
library(vegan) ## *must* be given before using the function
data(varespec)
mod <- GO2(varespec, k=2, far=5, tot=100, family="binomial", iterlim=1000)
plot(mod, label=TRUE)
ordiplot(mod, type="t")
ordiplot(mod, dis="si", type="t")
anova(mod)
mod1 <- update(mod, k=1)
anova(mod1, mod)
spanodev(mod1)
spanodev(mod1, mod)
}
\seealso{
\code{\link[VGAM]{cgo}} in \pkg{VGAM} package.
}

